chore: adding test folder and moving test to there
